DataUploader/requirements.txt
pydantic==1.10.2
python-dotenv==0.21.0
python-dateutil==2.8.2